+++ src/Params/GetInputParam.php
@@ -11,9 +11,7 @@ class GetInputParam extends InputParam
         if (!filter_has_var(INPUT_GET, $this->key) && isset($_GET[$this->key])) {
             return $_GET[$this->key];
         }
-        if ($this->isMulti()) {
-            return filter_input(INPUT_GET, $this->key, FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
-        }
-        return filter_input(INPUT_GET, $this->key);
+        $value = $this->isMulti() ? filter_input(INPUT_GET, $this->key, FILTER_DEFAULT, FILTER_REQUIRE_ARRAY) : filter_input(INPUT_GET, $this->key);
+        return $value !== null ? $value : $this->default;
     }
 }