+++ src/Params/InputParam.php
@@ -4,6 +4,8 @@ namespace Tomaj\NetteApi\Params;
 
 use Nette\Application\UI\Form;
 use Nette\Forms\Controls\BaseControl;
+use Nette\Utils\Html;
+use Tracy\Debugger;
 
 abstract class InputParam implements ParamInterface
 {
@@ -13,7 +15,7 @@ abstract class InputParam implements ParamInterface
     const TYPE_FILE      = 'FILE';
     const TYPE_COOKIE    = 'COOKIE';
     const TYPE_POST_RAW  = 'POST_RAW';
-    const TYPE_POST_JSON  = 'POST_JSON';
+    const TYPE_POST_JSON = 'POST_JSON';
 
     const OPTIONAL = false;
     const REQUIRED = true;
@@ -32,6 +34,15 @@ abstract class InputParam implements ParamInterface
 
     /** @var bool */
     protected $multi = false;
+
+    /** @var string */
+    protected $description = '';
+
+    /** @var mixed */
+    protected $default;
+
+    /** @var mixed */
+    protected $example;
 
     protected $errors = [];
 
@@ -83,6 +94,53 @@ abstract class InputParam implements ParamInterface
         return $this->multi;
     }
 
+    public function setDescription(string $description): self
+    {
+        $this->description = $description;
+        return $this;
+    }
+
+    public function getDescription(): string
+    {
+        return $this->description;
+    }
+
+    /**
+     * @param mixed $default
+     * @return self
+     */
+    public function setDefault($default): self
+    {
+        $this->default = $default;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getDefault()
+    {
+        return $this->default;
+    }
+
+    /**
+     * @param mixed $example
+     * @return self
+     */
+    public function setExample($example): self
+    {
+        $this->example = $example;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getExample()
+    {
+        return $this->example;
+    }
+
     public function updateConsoleForm(Form $form): void
     {
         $count = $this->isMulti() ? 5 : 1;  // TODO moznost nastavit kolko inputov sa ma vygenerovat v konzole, default moze byt 5
@@ -91,16 +149,23 @@ abstract class InputParam implements ParamInterface
             if ($this->isMulti()) {
                 $key = $key . '___' . $i;
             }
-            $this->addFormInput($form, $key);
+            $input = $this->addFormInput($form, $key);
+            if ($this->description) {
+                $input->setOption('description', Html::el('div', ['class' => 'param-description'])->setHtml($this->description));
+            }
+            if ($this->getExample() || $this->getDefault()) {
+                $default = $this->getExample() ?: $this->getDefault();
+                $default = is_array($default) ? ($default[$i] ?? null) : $default;
+                $input->setDefaultValue($default);
+            }
         }
     }
 
     protected function addFormInput(Form $form, string $key): BaseControl
     {
         if ($this->getAvailableValues()) {
-            $select = $form->addSelect($key, $this->getParamLabel(), array_combine($this->getAvailableValues(), $this->getAvailableValues()))
+            return $form->addSelect($key, $this->getParamLabel(), array_combine($this->getAvailableValues(), $this->getAvailableValues()))
                 ->setPrompt('Select ' . $this->getLabel());
-            return $select;
         }
         return $form->addText($key, $this->getParamLabel());
     }