+++ src/Params/JsonInputParam.php
@@ -6,7 +6,6 @@ use Exception;
 use JsonSchema\Validator;
 use Nette\Application\UI\Form;
 use Nette\Forms\Controls\BaseControl;
-use Nette\Utils\Html;
 
 class JsonInputParam extends InputParam
 {
@@ -30,7 +29,7 @@ class JsonInputParam extends InputParam
 
     public function getValue()
     {
-        $input = file_get_contents("php://input");
+        $input = file_get_contents("php://input") ?: $this->default;
         return json_decode($input, true);
     }
 
@@ -58,14 +57,12 @@ class JsonInputParam extends InputParam
 
     protected function addFormInput(Form $form, string $key): BaseControl
     {
-        return $form->addTextArea('post_raw', $this->getParamLabel())
-            ->setHtmlAttribute('rows', 10)
-            ->setOption('description', Html::el()->setHtml(
-                '<div id="show_schema_link"><a href="#" onclick="document.getElementById(\'json_schema\').style.display = \'block\'; document.getElementById(\'show_schema_link\').style.display = \'none\'; return false;">Show schema</a></div>
+        $this->description .= '<div id="show_schema_link"><a href="#" onclick="document.getElementById(\'json_schema\').style.display = \'block\'; document.getElementById(\'show_schema_link\').style.display = \'none\'; return false;">Show schema</a></div>
                             <div id="json_schema" style="display: none;">
                             <div><a href="#" onclick="document.getElementById(\'show_schema_link\').style.display = \'block\'; document.getElementById(\'json_schema\').style.display = \'none\'; return false;">Hide schema</a></div>'
-                . nl2br(str_replace(' ', '&nbsp;', json_encode(json_decode($this->schema), JSON_PRETTY_PRINT)))
-                . '</div>'
-            ));
+            . nl2br(str_replace(' ', '&nbsp;', json_encode(json_decode($this->schema), JSON_PRETTY_PRINT))) . '</div>';
+
+        return $form->addTextArea('post_raw', $this->getParamLabel())
+            ->setHtmlAttribute('rows', 10);
     }
 }