+++ tests/Params/ConsoleParamsTest.php
@@ -29,7 +29,7 @@ class ConsoleParamsTest extends TestCase
     public function testUpdateFormPostInputs()
     {
         $inputParams = [
-            (new PostInputParam('mykey1'))->setRequired(),
+            (new PostInputParam('mykey1'))->setRequired()->setExample('Hello world'),
             (new PostInputParam('mykey2'))->setMulti(),
             (new PostInputParam('mykey3'))->setRequired()->setAvailableValues(['a', 'b']),
         ];